+++ src/main/java/tornadofx/Properties.kt
@@ -1,9 +1,13 @@
 package tornadofx
 
+import javafx.beans.Observable
+import javafx.beans.binding.*
 import javafx.beans.property.*
 import javafx.beans.value.*
 import java.lang.reflect.Field
 import java.lang.reflect.Method
+import java.util.*
+import java.util.concurrent.Callable
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.*
 
@@ -211,3 +215,24 @@ operator fun ObservableBooleanValue.getValue(thisRef: Any?, property: KProperty<
 operator fun BooleanProperty.setValue(thisRef: Any?, property: KProperty<*>, value: Boolean) {
     this.value = value
 }
+
+fun <T> Property<T>.integerBinding(vararg dependencies: Observable, op: (T.() -> Int)): IntegerBinding
+        = Bindings.createIntegerBinding(Callable { op(value) }, this, *dependencies)
+
+fun <T> Property<T>.longBinding(vararg dependencies: Observable, op: (T?.() -> Long?)): LongBinding
+        = Bindings.createLongBinding(Callable { op(value) }, this, *dependencies)
+
+fun <T> Property<T>.doubleBinding(vararg dependencies: Observable, op: (T?.() -> Double?)): DoubleBinding
+        = Bindings.createDoubleBinding(Callable { op(value) }, this, *dependencies)
+
+fun <T> Property<T>.floatBinding(vararg dependencies: Observable, op: (T?.() -> Float?)): FloatBinding
+        = Bindings.createFloatBinding(Callable { op(value) }, this, *dependencies)
+
+fun <T> Property<T>.booleanBinding(vararg dependencies: Observable, op: (T?.() -> Boolean?)): BooleanBinding
+        = Bindings.createBooleanBinding(Callable { op(value) }, this, *dependencies)
+
+fun <T> Property<T>.stringBinding(vararg dependencies: Observable, op: (T?.() -> String?)): StringBinding
+        = Bindings.createStringBinding(Callable { op(value) }, this, *dependencies)
+
+fun <T, R> Property<T>.objectBinding(vararg dependencies: Observable, op: (T?.() -> R?)): Binding<R?>
+        = Bindings.createObjectBinding(Callable { op(value) }, this, *dependencies)